+++ aikau/src/test/resources/testApp/WEB-INF/classes/alfresco/site-webscripts/alfresco/forms/controls/Validation.get.js
@@ -13,8 +13,7 @@ model.jsonModel = {
       },
       "alfresco/services/DialogService",
       "alfresco/services/SiteService",
-      "aikauTesting/mockservices/FormControlValidationTestService",
-      "alfresco/services/ErrorReporter"
+      "aikauTesting/mockservices/FormControlValidationTestService"
    ],
    widgets: [
       {
@@ -125,6 +124,24 @@ model.jsonModel = {
                   }
                },
                {
+                  id: "SCOPED_TOPIC_VALIDATION",
+                  name: "alfresco/forms/controls/TextBox",
+                  config: {
+                     fieldId: "SCOPED_TOPIC_VALIDATION",
+                     label: "Validation Topic (Scoped response)",
+                     description: "Enter #fail as value to put into error state",
+                     value: "",
+                     validationConfig: [
+                        {
+                           scopeValidation: true,
+                           validation: "validationTopic",
+                           validationTopic: "ALF_VALIDATE_TOPIC_SCOPED_TEST",
+                           errorMessage: "Value should not be #fail"
+                        }
+                     ]
+                  }
+               },
+               {
                   id: "CUSTOMIZED_TOPIC_VALIDATION",
                   name: "alfresco/forms/controls/TextBox",
                   config: {
@@ -134,6 +151,7 @@ model.jsonModel = {
                      value: "test",
                      validationConfig: [
                         {
+                           warnOnly: true,
                            validation: "validationTopic",
                            validationTopic: "ALF_VALIDATE_SITE_IDENTIFIER",
                            validationValueProperty: "title",